academic_notes_ai/requirements.txt
langchain
langchain-groq
langchain-huggingface
langchain-community
sentence-transformers
faiss-cpu
python-dotenv